Async version of a feature registry
